Refactoring & documentation
