+++ Source/ModifyRequestNetworkService.swift
@@ -33,7 +33,8 @@ import Dispatch
 `ModifyRequestNetworkService` can be composed with a networkService to modify all outgoing requests.
 One could add auth tokens or API keys for specifics URLs.
  
-  - seealso: `NetworkServiceProviding`
+ - note: Requests can only be modified syncronously.
+ - seealso: `NetworkServiceProviding`
  */
 public final class ModifyRequestNetworkService: NetworkServiceProviding {
     
@@ -44,7 +45,7 @@ public final class ModifyRequestNetworkService: NetworkServiceProviding {
     ///
     /// - Parameters:
     ///   - networkService: a networkservice.
-    ///   - requestModifications: array of modifications to modify requests.
+    ///   - requestModifications: array of modifications to modify a requests.
     public init(networkService: NetworkServiceProviding, requestModifications: Array<(URLRequestConvertible) -> URLRequestConvertible>) {
         self.networkService = networkService
         self.requestModifications = requestModifications
@@ -58,80 +59,5 @@ public final class ModifyRequestNetworkService: NetworkServiceProviding {
         })
         let newResource = Resource(request: request, parse: resource.parse)
         return networkService.request(queue: queue, resource: newResource, onCompletionWithResponse: onCompletionWithResponse, onError: onError)
-    }
-}
-
-public extension URLRequestConvertible {
-    
-    /// Creates a new `URLRequestConvertible` with HTTPHeaderFields added into the new request.
-    /// Keep in mind that this overrides header fields which are already contained.
-    ///
-    /// - Parameter HTTPHeaderFields: the header fileds to add to the request
-    /// - Returns: a new `URLRequestConvertible`
-    func added(HTTPHeaderFields: [String: String]) -> URLRequestConvertible {
-        var request = asURLRequest()
-        let headerFiels = (request.allHTTPHeaderFields ?? [:]).merged(with: HTTPHeaderFields)
-        request.allHTTPHeaderFields = headerFiels
-        
-        return request
-    }
-    
-    /// Creates a new `URLRequestConvertible` with query items appended to the new request.
-    ///
-    /// - Parameter queryItems: the query items to append to the request
-    /// - Parameter overrideExisting: if true existing items with the same name will be overridden
-    /// - Returns: a new `URLRequestConvertible`
-    func appending(queryItems: [URLQueryItem], overrideExisting: Bool = true) -> URLRequestConvertible {
-        var request = asURLRequest()
-        guard let url = request.url else {
-            return self
-        }
-        request.url = url.appending(queryItems: queryItems)
-        return request
-    }
-    
-    /// Creates a new `URLRequestConvertible` with query parameters appended to the new request.
-    ///
-    /// - Parameter queryParameters: the parameters to append to the request
-    /// - Parameter overrideExisting: if true existing items with the same name will be overridden
-    /// - Returns: a new `URLRequestConvertible`
-    func appending(queryParameters: [String: String], overrideExisting: Bool = true) -> URLRequestConvertible {
-        return appending(queryItems: queryParameters.asURLQueryItems() )
-    }
-    
-    /// Creates a new `URLRequestConvertible` with all existing query items replaced with new ones.
-    ///
-    /// - Parameter queryItems: the queryItems to add to the request
-    /// - Returns: a new `URLRequestConvertible`
-    func replacingAllQueryItems(with queryItems: [URLQueryItem]) -> URLRequestConvertible {
-        var request = asURLRequest()
-        guard let url = request.url else {
-            return self
-        }
-        request.url = url.replacingAllQueryItems(with: queryItems)
-        return request
-    }
-    
-    /// Creates a new `URLRequestConvertible` with all existing query items replaced with new ones.
-    ///
-    /// - Parameter parameters: the parameters to add to the request
-    /// - Returns: a new `URLRequestConvertible`
-    func replacingAllQueryItems(with parameters: [String: String]) -> URLRequestConvertible {
-        return replacingAllQueryItems(with: parameters.asURLQueryItems() )
-    }
-}
-
-extension Dictionary {
-    /// Creates a new `Dictionary` with all key and their values merged. Keep in mind that this overrides all keys/values which are already contained.
-    ///
-    /// - Parameter HTTPHeaderFields: the header fileds to add to the request
-    /// - Returns: a new `NetworkRequestRepresening`
-    func merged(with dictionary: Dictionary<Key, Value>) -> Dictionary<Key, Value> {
-        var copySelf = self
-        for (key, value) in dictionary {
-            copySelf[key] = value
-        }
-        
-        return copySelf
     }
 }
